+++ src/app/(auth)/signUp/page.jsx
@@ -1,19 +1,18 @@
 "use client";
 
 import Container from "@/components/container/PageContainer";
-import Image from "next/image";
-import logo from "@/assets/img/img_logo.svg";
-import googleIcon from "@/assets/icon/ic_google.svg";
 import EmailInput from "@/components/input/EmailInput";
 import PasswordInput from "@/components/input/PasswordInput";
 import ConfirmPasswordInput from "@/components/input/ConfirmPasswordInput";
 import NicknameInput from "@/components/input/NicknameInput";
+import SubmitButton from "@/components/btn/auth/SubmitButton";
+import GoogleLoginButton from "@/components/btn/auth/GoogleLoginButton";
+import Logo from "@/layout/_components/Logo";
 import Link from "next/link";
-import { useRouter } from "next/navigation";
 import { useState } from "react";
 import { BASE_URL } from "@/constant/constant";
 
-const SignUpPage = () => {
+export default function SignUpPage() {
   const [isSignupSuccess, setIsSignupSuccess] = useState(false);
   const [isInputValid, setIsInputValid] = useState(false);
   const [values, setValues] = useState({
@@ -22,8 +21,6 @@ const SignUpPage = () => {
     password: "",
     confirmPassword: "",
   });
-
-  const router = useRouter();
 
   // 입력 필드 값 변경 시 실행되는 핸들러
   const handleChange = (e) => {
@@ -63,13 +60,7 @@ const SignUpPage = () => {
   return (
     <Container>
       <div className="mt-15 mb-30 flex flex-col items-center md:mt-30">
-        <Image
-          src={logo}
-          alt="Docthur 로고"
-          width={240}
-          height={54}
-          className="mb-10 md:h-18 md:w-80"
-        />
+        <Logo className="mb-10 h-[54px] w-60 md:h-18 md:w-80" />
         <form className="mb-[18px] space-y-6" onSubmit={handleSubmit}>
           <EmailInput
             value={values.email}
@@ -91,20 +82,9 @@ const SignUpPage = () => {
             onChange={handleChange}
             error={isInputValid}
           />
-          <button
-            type="submit"
-            className="h-12 w-[343px] rounded-xl bg-black font-semibold text-white md:w-[518px]"
-          >
-            회원가입
-          </button>
+          <SubmitButton type="회원가입" />
         </form>
-        <button
-          className="flex h-12 w-[343px] items-center justify-center gap-2 rounded-xl border border-gray-200 bg-white md:w-[518px]"
-          onClick={() => router.push("http://localhost:8080/auth/google")}
-        >
-          <Image src={googleIcon} alt="구글 아이콘" width={28} height={28} />
-          <p>Google로 시작하기</p>
-        </button>
+        <GoogleLoginButton />
         <div className="mt-6 space-x-2">
           <span>회원이신가요?</span>
           <Link href="/signIn" className="underline">
@@ -114,6 +94,4 @@ const SignUpPage = () => {
       </div>
     </Container>
   );
-};
-
-export default SignUpPage;
+}
