--- src/components/login/GoogleLoginButton.jsx
@@ -1,11 +1,0 @@
-import React from 'react'
-
-export function GoogleLoginButton() {
-  return (
-    <div className="w-85.5 h-12 bg-[#262626] flex justify-center items-center rounded-lg">
-      <div className='text-white font-semibold'>Google로 시작하기</div>
-    </div>
-  )
-}
-
-export default GoogleLoginButton
